move jacobian out of emulation into _shared
